% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryPoisson.R
\name{BinaryPoisson}
\alias{BinaryPoisson}
\title{A function to fit data with a mixed distribution of Poisson distribution using EM algorithm.}
\usage{
BinaryPoisson(x = c(162, 267, 271, 185, 111, 61, 27, 8, 3, 1))
}
\arguments{
\item{x}{An event (such as death) is arranged in increasing order of the number of occurrences per day within a year, and there is no number of days to complete the maximum number of occurrences with 0.}
}
\value{
an estimate of BinaryPoisson parameter (a,mu1,mu2)
}
\description{
A function to use a mixed distribution of Poisson distribution with mean mu1 and Poisson distribution with mean mu1 to fit data, which is known as the number of occurrences of a random event in unit time.And a is the percentage of the first distribution.
}
\examples{
\dontrun{
x=c(162,267,271,185,111,61,27,8,3,1)
BinaryPoisson(x)
}
}
