% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Metropolis}
\alias{Metropolis}
\title{A Metropolis sampler using Rcpp}
\usage{
Metropolis(sigma = 0.05, x0 = 25, N = 2000L)
}
\arguments{
\item{sigma}{the standard deviation of the normal distribution}

\item{x0}{the initial value}

\item{N}{the number of samples}
}
\value{
a random sample of size \code{N}
}
\description{
A Metropolis sampler using Rcpp
}
\examples{
\dontrun{
cpp.rw1=Metropolis(sigma=0.05,x0=25,N=2000)
plot(1:2000,cpp.rw1[,1],type='l',ylab="x",xlab='iteration',main='sd=0.05(Cpp)')
}
}
