% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scadest.R
\name{scadest}
\alias{scadest}
\title{A function to realize SCAD.(Jianqing Fan and Runze Li,2001)}
\usage{
scadest(x, y)
}
\arguments{
\item{x}{This is an n*d matrix which is also known as design matrix.}

\item{y}{This is an n*1 vector which is also known as response matrix}
}
\value{
an n*1 vector of the estimated parameter beta
}
\description{
Use scad method to solve parameter dimensionality reduction estimation in linear model.
}
\examples{
\dontrun{
data <- simulation1(n=40,sigma=1,beta=c(3, 1.5, 2, 0, 0, 0, 0, 0))
scadest(data$x, data$y)
}
}
