% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation1.R
\name{simulation1}
\alias{simulation1}
\title{Simulation model 1}
\usage{
simulation1(n = 40, sigma = 1, beta = c(3, 1.5, 2, 0, 0, 0, 0, 0))
}
\arguments{
\item{n}{sample size.}

\item{sigma}{the standard deviation of the error used in the simulation model.}

\item{beta}{the parameters needed to be estimated which is assumed in the simulation model.}
}
\value{
the data for similation
}
\description{
similar to the simulation model in SCAD (Jianqing Fan and Runze Li,2001)
}
\examples{
\dontrun{
data<-simulation1(n=40,sigma=1,beta=c(3, 1.5, 2, 0, 0, 0, 0, 0))
}
}
